+++ src/Domain/User/ValueObject/AuthResourceOwnerId.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Domain\User\ValueObject;
+
+use SixtyEightPublishers\ArchitectureBundle\Domain\ValueObject\AbstractStringValueObject;
+
+final class AuthResourceOwnerId extends AbstractStringValueObject
+{
+}