+++ src/Infrastructure/User/Doctrine/DbalType/AuthProviderCodeType.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Infrastructure\User\Doctrine\DbalType;
+
+use App\Domain\User\ValueObject\AuthProviderCode;
+use SixtyEightPublishers\ArchitectureBundle\Infrastructure\Doctrine\DbalType\AbstractStringValueObjectType;
+
+final class AuthProviderCodeType extends AbstractStringValueObjectType
+{
+    protected string $valueObjectClassname = AuthProviderCode::class;
+}