+++ src/Infrastructure/User/Doctrine/DbalType/AuthResourceOwnerIdType.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Infrastructure\User\Doctrine\DbalType;
+
+use App\Domain\User\ValueObject\AuthResourceOwnerId;
+use SixtyEightPublishers\ArchitectureBundle\Infrastructure\Doctrine\DbalType\AbstractStringValueObjectType;
+
+final class AuthResourceOwnerIdType extends AbstractStringValueObjectType
+{
+    protected string $valueObjectClassname = AuthResourceOwnerId::class;
+}