+++ src/Web/AdminModule/ApplicationModule/Control/AzureAuthSettingsForm/AzureAuthSettingsFormControlFactoryInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Web\AdminModule\ApplicationModule\Control\AzureAuthSettingsForm;
+
+interface AzureAuthSettingsFormControlFactoryInterface
+{
+    public function create(): AzureAuthSettingsFormControl;
+}